First steps to make Jetty act as a reverse proxy for the explorer
This adds a ProxyServlet instance that redirects requests from
/ewui/explorer/* to http://localhost:8082/*

It works for the HTML content. Unfortunately, Jetty does not support
WebSockets, so the shiny server isn't functional with just this. There are
reports of switching to wss: (WS+SSL) fixing the proxy problems, since the
encrypted stream gets passed end-to-end.

Note that all servlets get /ewui added to them currently. It seems not to be
included in the prefix setting.